added design smell detection and resolvement
